chore: Remove colorSelector component
