+++ apps/web/app/(app)/(survey-editor)/environments/[environmentId]/surveys/[surveyId]/edit/components/CardStylingSettings.tsx
@@ -12,7 +12,6 @@ import { TSurveyStyling, TSurveyType } from "@formbricks/types/surveys";
 import { Badge } from "@formbricks/ui/Badge";
 import { CardArrangementTabs } from "@formbricks/ui/CardArrangementTabs";
 import { ColorPicker } from "@formbricks/ui/ColorPicker";
-import { ColorSelector } from "@formbricks/ui/ColorSelector";
 import { FormControl, FormDescription, FormField, FormItem, FormLabel } from "@formbricks/ui/Form";
 import { Slider } from "@formbricks/ui/Slider";
 import { Switch } from "@formbricks/ui/Switch";
@@ -122,9 +121,10 @@ export const CardStylingSettings = ({
                 </div>
 
                 <FormControl>
-                  <ColorSelector
+                  <ColorPicker
                     color={field.value || COLOR_DEFAULTS.cardBackgroundColor}
-                    setColor={(color) => field.onChange(color)}
+                    onChange={(color) => field.onChange(color)}
+                    containerClass="max-w-xs"
                   />
                 </FormControl>
               </FormItem>
@@ -142,9 +142,10 @@ export const CardStylingSettings = ({
                 </div>
 
                 <FormControl>
-                  <ColorSelector
+                  <ColorPicker
                     color={field.value || COLOR_DEFAULTS.cardBorderColor}
-                    setColor={(color) => field.onChange(color)}
+                    onChange={(color) => field.onChange(color)}
+                    containerClass="max-w-xs"
                   />
                 </FormControl>
               </FormItem>
@@ -162,9 +163,10 @@ export const CardStylingSettings = ({
                 </div>
 
                 <FormControl>
-                  <ColorSelector
+                  <ColorPicker
                     color={field.value || COLOR_DEFAULTS.cardShadowColor}
-                    setColor={(color) => field.onChange(color)}
+                    onChange={(color) => field.onChange(color)}
+                    containerClass="max-w-xs"
                   />
                 </FormControl>
               </FormItem>
